check mem leaks

+++ parsing.h
@@ -6,7 +6,7 @@
 /*   By: cmrabet <cmrabet@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/08/30 02:01:28 by smokashi          #+#    #+#             */
-/*   Updated: 2023/11/06 09:16:05 by cmrabet          ###   ########.fr       */
+/*   Updated: 2023/11/07 16:54:17 by cmrabet          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -49,7 +49,7 @@ typedef struct s_shell
 	int					fail;
 	int					fd_tmp;
 	char				*pwd_tmp;
-	unsigned int		exit_code;
+	long long			exit_code;
 	char				**env_joind;
 	struct s_env		*env;
 	struct s_command	*command;
